Updated build config

+++ sandbox/clean.bat
@@ -1,2 +1,2 @@
-del /S *.hlt
-del /S *.log
+IF EXIST *.hlt del /S *.hlt
+IF EXIST *.log del /S *.log
